Rewritten slate plugin structure!
