+++ src/components/SlateEditor/plugins/saveForm/index.ts
@@ -1,0 +1,20 @@
+import { KeyboardEvent } from 'react';
+import { Editor } from 'new-slate';
+import isHotkey from 'is-hotkey';
+
+const isSaveHotkey = isHotkey('mod+s');
+
+export const savePlugin = (handleSubmit: () => void) => (editor: Editor) => {
+  const { onKeyDown: nextOnKeyDown } = editor;
+  editor.onKeyDown = (e: KeyboardEvent<HTMLDivElement>) => {
+    if (isSaveHotkey(e)) {
+      e.preventDefault();
+      handleSubmit();
+    } else if (nextOnKeyDown) {
+      nextOnKeyDown(e);
+    }
+  };
+  return editor;
+};
+
+export default savePlugin;
